requirements.txt
flask==1.1.2
numpy==1.18.1
opencv-python>=3.4.1.15
torch==1.4.0
torchvision==0.5.0
pillow==7.2.0
facenet-pytorch==2.4.1
pafy==0.5.5
requests==2.24.0
youtube-dl
efficientnet_pytorch==0.6.3
